make an index for each "blog"
The display is the same for all blogs. The default display contains
only the nav bar in the header and can be used for projects ; the page
layout adds the two columns with one for the presentation of Arthur and
one for the content ; the blog-index layout adds the latest posts of
the current blog.